Rebase endpointsec with master (#381)
* rename Docs -> docs (#372)

* Create CNAME

* Set theme jekyll-theme-cayman

* SantaGUI: ensure bundle listeners are invalidated (#373)

When setting a new bundle service listener, it was possible for an existing listener to be replaced without invalidating it first. This can cause crashes if a process somehow tries to connect to that listener later on.

* santad: Add /usr/lib/dyld to critical system binaries (#376)

dyld is also authorized by santad and a bad cache eviction plus trustd/ocspd not running can result in deadlock.

Fixes #375, probably.

* fix typo (inital -> intial) (#378)

* Add SNTXPCProxy and stubs for other components (#374)

This is the service interface for the XPC service responsible for vending connections to other components. This is necessary when running as a systemextension as we don't control the lifecycle of the daemon and so can't vend an XPC MachService directly

* Move Xcode project (#379)

* Create Xcode project

* sp

* keep bazel building

* use xcode11